Refactor generate_data_{objects,bundles}
